fix upload image

+++ src/pages/account/index.js
@@ -63,7 +63,7 @@ Page({
 
       this.setData({ 
         user: auth.user, 
-        avatar: image
+        avatar: image.url
       });
 
     } catch (err) {
@@ -84,44 +84,60 @@ Page({
     this.data.pages[index].func();
   },
   async onDeleteImage() {
-    try {
-      const mess = await deleteImageAPI(await havingToken());
-      if (mess.success) {
-        this.setData({ avatar: null });
-        this.onHide();
-      } else throw new Error("Server Error");
-    } catch (err) {
-      handleError(err.messega);
-      my.alert({
-        title: "Xoá thất bại",
-        success: () => {
-          this.onHide();
+    my.confirm({
+      title: 'Xác nhận xoá ảnh đại diện',
+      confirmButtonText: 'Xoá',
+      cancelButtonText: 'Huỷ',
+      success: async (result) => {
+        if (result.confirm) {
+          try {
+            const mess = await deleteImageAPI(await havingToken());
+            if (mess.success) {
+              this.setData({ avatar: null });
+              my.alert({
+                title: "Đã xoá",
+                success: () => {
+                  this.onHide();
+                }
+              });
+            } else throw new Error("Server Error");
+          } catch (err) {
+            handleError(err.messega);
+            my.alert({
+              title: "Xoá thất bại",
+              success: () => {
+                this.onHide();
+              }
+            });
+          }
         }
-      });
-    }
+      },
+    })
   },
   onChooseImage() {
     my.chooseImage({
       success: async (res) => {
         const img = res.filePaths[0];
         try {
-          const data = await postUploadImageAPI(await havingToken({ token, img }));
+          const data = await postUploadImageAPI(await havingToken({ image: img }));
           if (data.success) {
             try {
               const image = await getImageAPI(await havingToken());
-            
+              console.log(image);
               my.alert({ title: 'Cập nhật ảnh thành công' });
               this.setData({ avatar: image.url });
-            } catch (error) {
+            } catch (err) {
+              console.log(err);
               handleError(err.messega);
               my.alert({ title: 'Cập nhật ảnh thất bại' });
             }
-           
+      
           } else {
             my.alert({ title: 'Cập nhật ảnh thất bại' });
           } 
         } catch (err) {
           handleError(err.messega);
+          my.alert({ title: 'Cập nhật ảnh thất bại' });
         }
       },
       fail: (e) => {
